requirements.txt
# --- Core DL (install torch per your CUDA; Colab already provides it) ---
torch>=2.2,<3.0
torchvision>=0.15,<0.20

# --- Detection (YOLOv8) ---
ultralytics>=8.3.220,<8.4

# --- Feature matching (LoFTR via Kornia) ---
kornia>=0.6.12,<0.8

# --- Image I/O / processing ---
opencv-python-headless>=4.8
scikit-image>=0.22
imageio>=2.31
rawpy>=0.18
Pillow>=10.0

# --- Numerics / utilities ---
numpy>=1.26
pandas>=2.0
scipy>=1.11
tqdm>=4.66
PyYAML>=6.0
matplotlib>=3.8
